+++ src/modules/osubot/canvas/stat.js
@@ -4,7 +4,7 @@ import fs from 'fs'
 import path from 'path'
 // Import local files
 import util from './_util'
-import { promisify, promisifyGM } from './_util'
+import { promisify, promisifyGM, cachepath, assetspath } from './_util'
 import { getAvatar } from './avatar'
 
 /**
@@ -15,11 +15,11 @@ import { getAvatar } from './avatar'
  */
 export default async function drawStat(stat, statPrev) {
     const uid = stat.user_id
-    const dest = `cache${path.sep}osubot${path.sep}stat${path.sep}${uid}.jpg`
-    const avatarDest = `cache${path.sep}osubot${path.sep}avatar${path.sep}${uid}.jpg`
-    const avatarLargerDest = `cache${path.sep}osubot${path.sep}avatarl${path.sep}${uid}.jpg`
+    const dest = `${cachepath}/stat/${uid}.jpg`
+    const avatarDest = `${cachepath}/avatar/${uid}.jpg`
+    const avatarLargerDest = `${cachepath}/avatarl/${uid}.jpg`
     const ranks = ['XH', 'X', 'SH', 'S', 'A']
-    await promisify(fs.copyFile, `assets${path.sep}osubot${path.sep}image${path.sep}userbg${path.sep}c${Math.ceil(Math.random() * 5)}.jpg`, dest)
+    await promisify(fs.copyFile, `${assetspath}/image/userbg/c${Math.ceil(Math.random() * 5)}.jpg`, dest)
     await promisifyGM(
         gm(dest)
         .quality(100)
@@ -43,14 +43,14 @@ export default async function drawStat(stat, statPrev) {
         .drawRectangle(0, 530, 750, 1000)
         .drawRectangle(400, 50, 750, 160)
         .fill('#fff')
-        .font('assets/osubot/fonts/Exo2.0-Medium.otf')
+        .font(`${assetspath}/fonts/Exo2.0-Medium.otf`)
         .fontSize(40)
         .drawText(420, 100, stat.username)
         .fontSize(45)
         .drawText(30, 700, util.scorify(stat.playcount))
         .drawText(30, 790, stat.accuracy.slice(0, 3 + stat.accuracy.split('.')[0].length) + '%')
         .drawText(30, 880, util.scorify(stat.ranked_score))
-        .font('assets/osubot/fonts/Exo2.0-Bold.otf')
+        .font(`${assetspath}/fonts/Exo2.0-Bold.otf`)
         .drawText(30, 610, util.scorify(stat.pp_raw.split('.')[0]) + (stat.pp_raw.split('.')[1] ? ('.' + util.fillNumberReversed(stat.pp_raw.split('.')[1].slice(0, 2), 2)) : '') + 'pp')
         .fontSize(30)
         .drawText(60, 980, stat.count_rank_ssh)
@@ -58,7 +58,7 @@ export default async function drawStat(stat, statPrev) {
         .drawText(350, 980, stat.count_rank_sh)
         .drawText(495, 980, stat.count_rank_s)
         .drawText(640, 980, stat.count_rank_a)
-        .font('assets/osubot/fonts/Exo2.0-MediumItalic.otf')
+        .font(`${assetspath}/fonts/Exo2.0-MediumItalic.otf`)
         .fill('#ddd')
         .drawText(420, 140, 'Lv. ' + parseInt(stat.level))
         .fontSize(25)
@@ -67,7 +67,7 @@ export default async function drawStat(stat, statPrev) {
         .drawText(30, 840, 'Ranked score')
         .drawText(30, 570, 'Performance points')
         .fill('#fff')
-        .font('assets/osubot/fonts/Venera-700.otf')
+        .font(`${assetspath}/fonts/Venera-700.otf`)
         .fontSize(45)
         .drawText(30, 480, '#' + util.scorify(stat.pp_rank))
         .gravity('NorthEast')
@@ -79,7 +79,7 @@ export default async function drawStat(stat, statPrev) {
         await promisifyGM(
             gm(dest)
             .quality(100)
-            .font('assets/osubot/fonts/Exo2.0-Regular.otf')
+            .font(`${assetspath}/fonts/Exo2.0-Regular.otf`)
             .fill('#ddd')
             .fontSize(30)
             .drawText(30, 440, (-parseInt(diff.pp_rank) >= 0 ? '+' : '') + (-diff.pp_rank))
@@ -101,14 +101,14 @@ export default async function drawStat(stat, statPrev) {
         await promisifyGM(
             gm(dest)
             .quality(100)
-            .composite(`assets/osubot/image/rank/${ranks[rank]}.png`)
+            .composite(`${assetspath}/image/rank/${ranks[rank]}.png`)
             .geometry(`+${30 + rank * 145}+860`)
         )
     await promisifyGM(
         gm(dest)
         .quality(100)
         .gravity('NorthEast')
-        .composite('assets/osubot/image/flags/' + stat.country + '.png')
+        .composite(`${assetspath}/image/flags/${stat.country}.png`)
         .geometry('+30+455')
     )
     return 'file://' + process.cwd() + path.sep + dest