+++ src/modules/osubot/commands/bp.js
@@ -45,12 +45,20 @@ export default {
             }
             const path = await drawBest(best, map, status)
             if (path)
-                msg.send([{
-                    type: 'image',
-                    data: {
-                        file: path,
-                    }
-                }])
+                msg.send([
+                    {
+                        type: 'image',
+                        data: {
+                            file: path,
+                        }
+                    },
+                    {
+                        type: 'text',
+                        data: {
+                            text: `\nhttps://osu.ppy.sh/b/${best.beatmap_id}`
+                        }
+                    },
+                ])
             else msg.send(`osubot: bp: ${MESSAGES.QUERY_CANVAS_FAIL}`)
         } catch (err) {
             msg.error(err)