+++ src/modules/osubot/init.js
@@ -1,17 +1,18 @@
 import fs from 'fs'
 import { statdb } from './db'
-import { mod } from '../../core/log'
+import { modLog } from '../../core/log'
+import { cachepath } from './canvas/_util'
 
 const initPaths = [
-    'cache/osubot',
-    'cache/osubot/avatar',
-    'cache/osubot/avatarl',
-    'cache/osubot/recent',
-    'cache/osubot/recentbg',
-    'cache/osubot/stat',
-    'cache/osubot/mapbg',
-    'cache/osubot/best',
-    'cache/osubot/mapfile'
+    `${cachepath}`,
+    `${cachepath}/avatar`,
+    `${cachepath}/avatarl`,
+    `${cachepath}/recent`,
+    `${cachepath}/recentbg`,
+    `${cachepath}/stat`,
+    `${cachepath}/mapbg`,
+    `${cachepath}/best`,
+    `${cachepath}/mapfile`
 ]
 
 /**
@@ -24,10 +25,10 @@ export default async function refresher() {
     const time = new Date(`${new Date().getFullYear()}-${new Date().getMonth() + 1}-${new Date().getDate() + 1}`)
     setTimeout(async () => {
         statdb.refreshAllStat()
-              .then(() => mod('osubot initializer', 'Refreshed user status'))
+              .then(() => modLog('osubot initializer', 'Refreshed user status'))
         setInterval(async () => {
             statdb.refreshAllStat()
-                  .then(() => mod('osubot initializer', 'Refreshed user status'))
+                  .then(() => modLog('osubot initializer', 'Refreshed user status'))
         }, 86400000)
     }, time.getTime() - Date.now())
 }